+++ src/scripts/index.ts
@@ -1,2 +1,4 @@
 export { setupDemo } from './setupDemo'
 export { createCoreInstance, Environment } from './createCoreInstance'
+export { sendBalanceToProviderAccount } from './sendBalanceToProviderAccount'
+export { ISetup, deployAllContracts, setupContracts } from './setupContracts'
