+++ src/tests/Cache.test.ts
@@ -1,9 +1,8 @@
-import { createDbConnection } from '../storage/createDbConnection'
-import { deleteDatabase, resetDatabase } from './utils'
 import { Connection, Repository } from 'typeorm'
-import { ScheduledTransaction, EMetatransactionState } from '../entities'
-import { Cache } from '../storage'
 import { addMinutes } from 'date-fns'
+import { ScheduledTransaction, EMetatransactionState } from '../entities'
+import { Cache, createDbConnection } from '../storage'
+import { deleteDatabase, resetDatabase } from './utils'
 
 jest.setTimeout(7000)
