+++ src/tests/Collector.test.ts
@@ -1,10 +1,9 @@
-import { addMinutes } from 'date-fns'
-import { createDbConnection } from '../storage/createDbConnection'
-import { deleteDatabase, resetDatabase } from './utils'
 import { Connection, Repository } from 'typeorm'
-import { ScheduledTransaction, IMetatransaction, EMetatransactionState } from '../entities'
-import { Cache } from '../storage'
+import { addMinutes } from 'date-fns'
 import { Collector } from '../model'
+import { Cache, createDbConnection } from '../storage'
+import { deleteDatabase, resetDatabase } from './utils'
+import { ScheduledTransaction, IMetatransaction, EMetatransactionState } from '../entities'
 
 jest.setTimeout(17000)
