+++ src/tests/Executor.test.ts
@@ -2,10 +2,10 @@ import Web3 from 'web3'
 import { Executor } from '../model'
 import { addMinutes } from 'date-fns'
 import { time } from '@openzeppelin/test-helpers'
-import { deployAllContracts, getAccounts, ISetup, setupContracts } from './setupContracts'
-import { sendBalanceToProviderAccount } from './sendBalanceToProviderAccount'
+import { deployAllContracts, getAccounts, ISetup, setupContracts } from '../scripts/setupContracts'
+import { sendBalanceToProviderAccount } from '../scripts/sendBalanceToProviderAccount'
 import { BLOCKCHAIN_HTTP_URL, MNEMONIC_PHRASE } from './constants'
-import ERC677Data from './contracts/ERC677.json'
+import ERC677Data from '../scripts/contracts/ERC677.json'
 import { AbiItem } from 'web3-utils'
 import { BlockchainDate } from '../time'
