+++ src/tests/ListenerRunner.ts
@@ -1,6 +1,6 @@
 import Web3 from 'web3'
 import { addMinutes } from 'date-fns'
-import { deployAllContracts, ISetup, setupContracts } from './setupContracts'
+import { deployAllContracts, ISetup, setupContracts } from '../scripts/setupContracts'
 import { BLOCKCHAIN_HTTP_URL } from './constants'
 import { BlockchainDate } from '../time'
 import { EListenerEvents, IListener } from '../model'
